[C++] Fix wrong type for Unbind.

+++ lib/v2.0/C++/pprzlink/IvyLink.cpp
@@ -89,7 +89,7 @@ namespace pprzlink {
     return id;
   }
 
-  void IvyLink::UnbindMessage(int bindId)
+  void IvyLink::UnbindMessage(long bindId)
   {
     if (messagesCallbackMap.find(bindId) != messagesCallbackMap.end()) {
       bus->UnbindMsg(bindId);
